+++ EnrollmentSystem/sectionmenu.cs
@@ -34,17 +34,13 @@ namespace EnrollmentSystem
             }
             Currrcb.Items.AddRange(arrayList.ToArray());
             arrayList.Clear();
-            foreach (DataRow dset in checker.FillCourse().Rows)
-            {
-                arrayList.Add(string.Join(";", dset.ItemArray.Select(item => item.ToString())));
-            }
-            coursecb.Items.AddRange(arrayList.ToArray());
-            arrayList.Clear();
-            string[] yls = { "First Year", "Second  Year", "Third Year", "Fourth Year" };
-            ylcb.Items.AddRange(yls);
-
-            string[] sems = { "First Sem", "Second Sem" };
-            semcb.Items.AddRange(sems);
+            //foreach (DataRow dset in checker.FillCourse().Rows)
+            //{
+            //    arrayList.Add(string.Join(";", dset.ItemArray.Select(item => item.ToString())));
+            //}
+            //coursecb.Items.AddRange(arrayList.ToArray());
+            //arrayList.Clear();
+            
             f.enableShow(clearbtn);
             f.disableHide(deletebtn);
             f.enableShow(addbtn);
@@ -75,12 +71,12 @@ namespace EnrollmentSystem
                 if (sectioncode != null)
                 {
                     checker.AddSection(sectioncode, Currrcb.SelectedItem.ToString(), coursecb.SelectedItem.ToString(), ylcb.SelectedItem.ToString(), semcb.SelectedItem.ToString(), 0) ;
-                    MessageBox.Show("Section created successfully.", "Section Created");
+                    MessageBox.Show("Section created successfully.", "Section Created", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     clearData();
                 }
                 else
                 {
-                    MessageBox.Show("Please check all the information you entered.", "Missing Information");
+                    MessageBox.Show("Please check all the information you entered.", "Missing Information", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
              }
 
@@ -97,14 +93,24 @@ namespace EnrollmentSystem
 
         private void deletebtn_Click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("Do you want to delete the section '" + sectiontxt.Text + "' ?", "Delete Section?", MessageBoxButtons.YesNo);
+            DialogResult result = MessageBox.Show("Do you want to delete the section '" + sectiontxt.Text + "'?\n" +
+                "Schedules under this section will also be deleted.", "Delete Section?", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
             if (result == DialogResult.Yes)
             {
                 try
                 {
-                    checker.DeleteSection(sectiontxt.Text);
-                    MessageBox.Show("Section deleted successfully.", "Section Deleted");
-                    clearData();
+                    if (numtxt.Text != "0")
+                    {
+                        MessageBox.Show("Cannot delete settion because there are students added in the section", "Cannot Delete Section", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        checker.DeleteSection(sectiontxt.Text);
+                        checker.DeleteSchedBySection(sectiontxt.Text);
+                        checker.DeleteSchedBySectionIrregular(sectiontxt.Text);
+                        MessageBox.Show("Section deleted successfully.", "Section Deleted", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        clearData();
+                    }
                 }
                 catch (Exception ex)
                 {
@@ -170,17 +176,26 @@ namespace EnrollmentSystem
 
         private void Currrcb_SelectedIndexChanged(object sender, EventArgs e)
         {
-            letter = 'A';
-            if (Currrcb.SelectedItem != null && coursecb.SelectedItem != null && ylcb.SelectedItem != null && semcb.SelectedItem != null)
+            GenerateCode();
+            coursecb.Items.Clear();
+            coursecb.DropDownHeight = 106;
+            if (Currrcb.SelectedItem != null)
             {
-                while (checker.IfSectionExist(sectioncode = coursecb.SelectedItem.ToString() + (ylcb.SelectedIndex + 1) + "." + (semcb.SelectedIndex + 1) + letter))
+                try
                 {
-                    letter++;
+                    foreach (DataRow dset in checker.FillSelectedCourse(Currrcb.SelectedItem.ToString()).Rows)
+                    {
+                        arrayList.Add(string.Join(";", dset.ItemArray.Select(item => item.ToString())));
+                    }
+                    coursecb.Items.AddRange(arrayList.ToArray());
+                    arrayList.Clear();
                 }
-                sectioncode = coursecb.SelectedItem.ToString() + (ylcb.SelectedIndex + 1) + "." + (semcb.SelectedIndex + 1) + letter;
-                sectiontxt.Text = sectioncode;
+
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
             }
-            clearbtn.Enabled = true; 
         }
 
         private void Currrcb_KeyDown(object sender, KeyEventArgs e)
@@ -193,6 +208,48 @@ namespace EnrollmentSystem
             es = new editsection(sectiontxt.Text);
             es.ShowDialog();
             clearData();
+        }
+
+        private void coursecb_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            GenerateCode();
+
+            ylcb.Items.Clear();
+            semcb.Items.Clear();
+            ylcb.DropDownHeight = 106;
+            semcb.DropDownHeight = 106;
+
+            if (coursecb.SelectedItem != null)
+            {
+                try
+                {
+                    ylcb.Items.AddRange(f.ReturnYearLevel(checker.ReturnCourseYears(coursecb.SelectedItem.ToString())).ToArray());
+                    semcb.Items.AddRange(f.ReturnSems(checker.ReturnCourseSems(coursecb.SelectedItem.ToString())).ToArray());
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
+            }
+        }
+        public void GenerateCode()
+        {
+            letter = 'A';
+            if (Currrcb.SelectedItem != null && coursecb.SelectedItem != null && ylcb.SelectedItem != null && semcb.SelectedItem != null)
+            {
+                while (checker.IfSectionExist(sectioncode = coursecb.SelectedItem.ToString() + (ylcb.SelectedIndex + 1) + "." + (semcb.SelectedIndex + 1) + letter))
+                {
+                    letter++;
+                }
+                sectioncode = coursecb.SelectedItem.ToString() + (ylcb.SelectedIndex + 1) + "." + (semcb.SelectedIndex + 1) + letter;
+                sectiontxt.Text = sectioncode;
+            }
+            clearbtn.Enabled = true;
+        }
+
+        private void ylcb_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            GenerateCode();
         }
     }
 }
